Made dEdx back intoa float. It was accidentally left as a double while testing the recent changes to the HDDM organization.

+++ libraries/HDDM/hddm_s.h
@@ -267,7 +267,7 @@ typedef struct {
 
 typedef struct {
    float                E;
-   double               dEdx;
+   float                dEdx;
    float                dradius;
    bool_t               primary;
    int32_t              ptype;
